Tests/LibWeb/Text/expected/indexeddb-close-waits-for-transaction.txt

setup.upgradeneeded
setup.success
conn.close() called
put.success
tx.complete
upgrade.upgradeneeded
upgrade.success
DONE
